+++ test/integration/clean/test_clean.py
@@ -1,6 +1,7 @@
 import glob
 import itertools
 import re
+import os
 
 import pytest
 import tt_helper
@@ -11,7 +12,21 @@ skip_cluster_cond = utils.is_tarantool_less_3()
 skip_cluster_reason = "skip cluster instances test for Tarantool < 3"
 
 
+# Unchecked precondition for not force_mode:
+# more than 0 total files to remove.
 def check_clean(tt, tt_app, stop_targets, target, *args):
+    force_mode = "-f" in args or "--force" in args
+    verbose_mode = "-V" in args or "--verbose" in args
+    n_of_target_instances = len(tt_app.instances_of(target))
+    if not force_mode and n_of_target_instances > 1:
+        # Can't emulate more than one confirmation.
+        # See a note for go func AskConfirm().
+        pytest.skip("Misconfig: non force mode with multi instance app")
+    if force_mode:
+        input = ""
+    else:
+        input = "y\n" * n_of_target_instances
+
     # Stop the specified targets.
     for target in stop_targets:
         rc, _ = tt.exec("stop", target, "-y")
@@ -20,8 +35,21 @@ def check_clean(tt, tt_app, stop_targets, target, *args):
     # Store original state.
     orig_status = tt_helper.status(tt)
 
+    # Collect files to remove.
+    to_remove = {}
+    for inst in tt_app.instances:
+        if inst in tt_app.instances_of(target):
+            logs = tt_helper.log_files(tt_app, [inst])
+            snaps = tt_helper.snap_files(tt_app, [inst])
+            wals = tt_helper.wal_files(tt_app, [inst])
+            to_remove[inst] = {
+                "logs": glob.glob(logs[0]),
+                "snaps": glob.glob(snaps[0]),
+                "wals": glob.glob(wals[0]),
+            }
+
     # Do clean.
-    rc, out = tt.exec("clean", target, *args)
+    rc, out = tt.exec("clean", target, *args, input=input)
     assert rc == 0
 
     stop_instances = tt_app.instances_of(*stop_targets)
@@ -29,6 +57,7 @@ def check_clean(tt, tt_app, stop_targets, target, *args):
 
     # Check the instances.
     status = tt_helper.status(tt)
+    cleaned = 0
     for inst in tt_app.instances:
         was_running = inst in tt_app.running_instances
         assert status[inst]["STATUS"] == orig_status[inst]["STATUS"]
@@ -46,10 +75,34 @@ def check_clean(tt, tt_app, stop_targets, target, *args):
                     assert not glob.glob(tt_helper.log_files(tt_app, [inst])[0])
                     assert not glob.glob(tt_helper.snap_files(tt_app, [inst])[0])
                     assert not glob.glob(tt_helper.wal_files(tt_app, [inst])[0])
+                    cleaned += 1
+                    if not force_mode:
+                        n_logs = len(to_remove[inst]["logs"])
+                        if n_logs > 0:
+                            msg = r"to remove {} file\(s\) in \"{}\"\n".format(
+                                n_logs, tt_app.log_path(inst)
+                            )
+                            assert len(re.findall(msg, out)) == 1
+                        n_snaps_wals = len(to_remove[inst]["snaps"]) + \
+                                       len(to_remove[inst]["wals"])
+                        if n_snaps_wals > 0:
+                            msg = r"to remove {} file\(s\) in \"{}\"\n".format(
+                                n_snaps_wals, tt_app.lib_path(inst)
+                            )
+                            assert len(re.findall(msg, out)) == 1
+                    if not force_mode and verbose_mode:
+                        for dir in to_remove[inst]:
+                            for file in to_remove[inst][dir]:
+                                name = os.path.basename(file)
+                                msg = r"to remove \"\.\/{}\"\n".format(name)
+                                assert len(re.findall(msg, out)) == 1
                 else:
                     assert msg in out
             else:
                 assert f"{inst}...\t[ERR]" in out
+    if not force_mode:
+        assert len(re.findall(r"Total files to remove: \d+", out)) == cleaned
+        assert len(re.findall(r"Confirm \[y/n\]:", out)) == cleaned
 
 
 def post_start_clean_decorator(func):
@@ -68,6 +121,31 @@ def post_start_clean_decorator(func):
         )
 
     return wrapper_func
+
+
+################################################################
+# Simple app
+
+tt_simple_data_app = dict(
+    app_path="test_data_app.lua",
+    app_name="app",
+    post_start=tt_helper.post_start_base
+)
+
+
+# Ask confirmation with verbose mode variations.
+@pytest.mark.tt_app(**dict(tt_simple_data_app, running_targets=["app"]))
+@pytest.mark.parametrize(
+    "args",
+    [
+        pytest.param([], id="verbose_mode:none"),
+        pytest.param(["-V"], id="verbose_mode:short"),
+        pytest.param(["--verbose"], id="verbose_mode:long"),
+        pytest.param(["-V", "-f"], id="force_mode:verbose_ignored-no_confirmation"),
+    ],
+)
+def test_clean_single_inst(tt, tt_app, args):
+    check_clean(tt, tt_app, ["app"], "app", *args)
 
 
 ################################################################